Use console.log instead of throwing errors to make cake watch work

+++ Cakefile
@@ -40,9 +40,9 @@ compile = (browser) ->
   process_ghcr = ->
     ghcr_path = "#{build_path}/data/ghcr.coffee"
     fs.writeFile ghcr_path, ghcrContents.join('\n\n'), 'utf8', (err) ->
-      throw err if err
+      return console.log(stderr) if err
       exec "coffee --compile #{ghcr_path}", (err, stdout, stderr) ->
-        throw err if err
+        return console.log(stderr) if err
         out = stdout + stderr
         console.log out if out.length
         fs.unlink ghcr_path, ->
@@ -58,12 +58,13 @@ compile = (browser) ->
       command = switch extension
         when 'coffee' then "coffee --compile --output #{build_path}/lib #{file} "
         when 'scss', 'sass' then "sass --no-cache #{file} >> #{build_path}/data/ghcr.css && rm #{file}"
-        else throw "Unknown extension: #{file}"
-      exec command, (err, stdout, stderr) ->
-        throw err if err
-        out = stdout + stderr
-        console.log out if out.length
-        console.log "Compiled #{file}"
+        else console.log("\nUnknown extension: #{file}\n")
+      if command?
+        exec command, (err, stdout, stderr) ->
+          return console.log(stderr) if err
+          out = stdout + stderr
+          console.log out if out.length
+          console.log "Compiled #{file}"
 
   build_path = "builds/#{browser}"
   copy()
@@ -71,7 +72,7 @@ compile = (browser) ->
   ghcrContents  = []
   for file, index in ghcrFiles then do (file, index) ->
     fs.readFile "#{file}.coffee", 'utf8', (err, fileContents) ->
-      throw err if err
+      return console.log(stderr) if err
       ghcrContents[index] = fileContents
       process_ghcr() if ghcrContents.length == ghcrFiles.length
   compile_singles()
@@ -80,7 +81,7 @@ zip = (browser) ->
   build_path = "builds/#{browser}"
   exec "cd #{build_path}"
   exec "zip -r #{build_path} #{browser}", (err, stdout, stderr) ->
-    throw err if err
+    return console.log(stderr) if err
     out = stdout + stderr
     console.log out if out.length
     console.log "Zip #{build_path}.zip prepared"
